monitor/requirements.txt
httpx==0.22.0
kafka-python==2.0.2
-r requirements_dev.txt